requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
attrs==20.2.0             # via pytest
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
django-filter==2.4.0      # via -r requirements.in
django==3.1.2             # via -r requirements.in, django-filter, djangorestframework
djangorestframework==3.12.1  # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
importlib-metadata==2.0.0  # via pluggy, pytest
iniconfig==1.1.1          # via pytest
packaging==20.4           # via pytest
pillow==8.0.1             # via -r requirements.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.6    # via -r requirements.in
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==4.1.0      # via -r requirements.in
pytest==6.1.2             # via pytest-django
python-dotenv==0.15.0     # via -r requirements.in
pytz==2020.1              # via django
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via packaging
sqlparse==0.4.1           # via django
toml==0.10.1              # via pytest
urllib3==1.25.11          # via requests
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
